apps/api/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
garminconnect==0.2.4
